fixed WORD definition. a WORD is usually of the same size as a long. this also fixes: Cunlzexe.cpp:238: warning: comparison is always true due to limited range of data type, which of course was pretty serious!
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@1112 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/fileio/Cunlzexe.h
@@ -14,7 +14,7 @@
 
 #include <vector>
 
-typedef unsigned short WORD;
+typedef unsigned long WORD;
 typedef unsigned char BYTE;
 
 class Cunlzexe
